requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.1.0
google-genai>=0.2.0
pinecone-client>=3.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
